% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_methods.R
\docType{methods}
\name{constrain_variables}
\alias{constrain_variables}
\alias{constrain_variables,StanBase-method}
\alias{constrain_variables,StanBase,StanBase-method}
\title{Constrain a vector of variables.}
\usage{
constrain_variables(stan_object, unconstrained_variables)

\S4method{constrain_variables}{StanBase}(stan_object, unconstrained_variables)
}
\arguments{
\item{stan_object}{A \code{StanBase} object.}

\item{unconstrained_variables}{Vector of unconstrained variables.}
}
\description{
Constrain a vector of variables.
}
