% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_methods.R
\docType{methods}
\name{grad_log_prob}
\alias{grad_log_prob}
\alias{grad_log_prob,StanBase-method}
\alias{grad_log_prob,StanBase,StanBase-method}
\title{Calculate the log probability and its gradient of the
model given a vector of unconstrained variables.}
\usage{
grad_log_prob(stan_object, unconstrained_variables, jacobian = TRUE)

\S4method{grad_log_prob}{StanBase}(stan_object, unconstrained_variables, jacobian = TRUE)
}
\arguments{
\item{stan_object}{A \code{StanBase} object.}

\item{unconstrained_variables}{Vector of unconstrained variables.}

\item{jacobian}{Whether to include the Jacobian adjustment.}
}
\description{
Calculate the log probability and its gradient of the
model given a vector of unconstrained variables.
}
