% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_methods.R
\docType{methods}
\name{hessian}
\alias{hessian}
\alias{hessian,StanBase-method}
\alias{hessian,StanBase,StanBase-method}
\title{Calculate the log probability, its gradient, and its hessian matrix of the
model given a vector of unconstrained variables.}
\usage{
hessian(stan_object, unconstrained_variables, jacobian = TRUE)

\S4method{hessian}{StanBase}(stan_object, unconstrained_variables, jacobian = TRUE)
}
\arguments{
\item{stan_object}{A \code{StanBase} object.}

\item{unconstrained_variables}{Vector of unconstrained variables.}

\item{jacobian}{Whether to include the Jacobian adjustment.}
}
\description{
Calculate the log probability, its gradient, and its hessian matrix of the
model given a vector of unconstrained variables.
}
