% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_methods.R
\docType{methods}
\name{log_prob}
\alias{log_prob}
\alias{log_prob,StanBase-method}
\alias{log_prob,StanBase,StanBase-method}
\title{Calculate the log probability of the model given a vector of
unconstrained variables.}
\usage{
log_prob(stan_object, unconstrained_variables, jacobian = TRUE)

\S4method{log_prob}{StanBase}(stan_object, unconstrained_variables, jacobian = TRUE)
}
\arguments{
\item{stan_object}{A \code{StanBase} object.}

\item{unconstrained_variables}{Vector of unconstrained variables.}

\item{jacobian}{Whether to include the Jacobian adjustment.}
}
\description{
Calculate the log probability of the model given a vector of
unconstrained variables.
}
