% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_methods.R
\name{loo.StanBase}
\alias{loo.StanBase}
\alias{loo}
\title{Calculate approximate leave-one-out cross-validation (LOO-CV) for a model.}
\usage{
\method{loo}{StanBase}(x, pointwise_ll_fun, additional_args = list(), moment_match = FALSE, ...)
}
\arguments{
\item{x}{A \code{StanBase} object.}

\item{pointwise_ll_fun}{Function that calculates the pointwise log-likelihood
given a vector of parameter values.}

\item{additional_args}{List of additional arguments to be passed to
\code{pointwise_ll_fun}.}

\item{moment_match}{(logical) Whether to use a
\link[loo:loo_moment_match]{moment-matching} correction for problematic
observations.}

\item{...}{Additional arguments to be passed to \code{loo::loo()}.}
}
\description{
Calculate approximate leave-one-out cross-validation (LOO-CV) for a model.
}
