% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplace.R
\name{stan_laplace}
\alias{stan_laplace}
\title{stan_laplace}
\usage{
stan_laplace(
  fn,
  par_inits,
  ...,
  grad_fun = NULL,
  lower = -Inf,
  upper = Inf,
  seed = NULL,
  refresh = NULL,
  output_dir = NULL,
  output_basename = NULL,
  sig_figs = NULL,
  mode = NULL,
  jacobian = NULL,
  draws = NULL,
  opt_args = NULL
)
}
\arguments{
\item{fn}{Function to estimate parameters for}

\item{par_inits}{Initial values}

\item{...}{Additional arguments to pass to the function}

\item{grad_fun}{Function calculating gradients w.r.t. each parameter}

\item{lower}{Lower bound constraint(s) for parameters}

\item{upper}{Upper bound constraint(s) for parameters}

\item{seed}{Random seed}

\item{refresh}{Number of iterations for printing}

\item{output_dir}{Directory to store outputs}

\item{output_basename}{Basename to use for output files}

\item{sig_figs}{Number of significant digits to use for printing}

\item{mode}{Mode for the laplace approximation, can either be a vector of
values, a StanOptimize object, or NULL.}

\item{jacobian}{(logical) Whether or not to use the Jacobian adjustment for
constrained variables.}

\item{draws}{(positive integer) Number of approximate posterior
samples to draw and save.}

\item{opt_args}{(named list) A named list of optional arguments to pass to
\code{stan_optimize()} if \code{mode=NULL}.}
}
\value{
\code{StanLaplace} object
}
\description{
Estimate parameters using Stan's laplace algorithm
}
