% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{stan_sample}
\alias{stan_sample}
\title{stan_sample}
\usage{
stan_sample(
  fn,
  par_inits,
  ...,
  algorithm = "hmc",
  engine = "nuts",
  num_chains = 4,
  num_samples = 1000,
  num_warmup = 1000,
  save_warmup = FALSE,
  thin = 1,
  output_dir = tempdir(),
  control = list()
)
}
\description{
Estimate parameters using Stan's sampling algorithms
}
