% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variational.R
\name{stan_variational}
\alias{stan_variational}
\title{stan_variational}
\usage{
stan_variational(
  fn,
  par_inits,
  ...,
  algorithm = "meanfield",
  output_samples = 1000,
  iter = 1000,
  grad_samples = 1,
  elbo_samples = 100,
  eval_elbo = 100,
  output_dir = tempdir(),
  control = list()
)
}
\description{
Estimate parameters using Stan's variational inference algorithms
}
