% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StanEstimators-package.R
\docType{package}
\name{StanEstimators-package}
\alias{StanEstimators-package}
\alias{StanEstimators}
\title{The StanEstimators package.}
\description{
ToDO
}
\author{
\strong{Maintainer}: Andrew R. Johnson \email{andrew.johnson@arjohnsonau.com} (\href{https://orcid.org/0000-0001-7000-8065}{ORCID})

Other contributors:
\itemize{
  \item Jonah Gabry \email{jsg2201@columbia.edu} [contributor]
  \item Rok Češnovar \email{rok.cesnovar@fri.uni-lj.si} [contributor]
  \item Stan Development Team (CmdStan sources and headers) [copyright holder]
  \item Lawrence Livermore National Security (SUNDIALS sources and headers) [copyright holder]
}

}
