% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplace.R
\docType{methods}
\name{summary-StanLaplace}
\alias{summary-StanLaplace}
\alias{summary,StanLaplace-method}
\title{Summary method for objects of class \code{StanLaplace}.}
\usage{
\S4method{summary}{StanLaplace}(object, ...)
}
\arguments{
\item{object}{A \code{StanLaplace} object.}

\item{...}{Additional arguments, currently unused.}
}
\description{
Summary method for objects of class \code{StanLaplace}.
}
