% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\docType{methods}
\name{summary-StanMCMC}
\alias{summary-StanMCMC}
\alias{summary,StanMCMC-method}
\title{Summary method for objects of class \code{StanMCMC}.}
\usage{
\S4method{summary}{StanMCMC}(object, ...)
}
\arguments{
\item{object}{A \code{StanMCMC} object.}

\item{...}{Additional arguments, currently unused.}
}
\description{
Summary method for objects of class \code{StanMCMC}.
}
