% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\docType{methods}
\name{summary-StanOptimize}
\alias{summary-StanOptimize}
\alias{summary,StanOptimize-method}
\title{Summary method for objects of class \code{StanOptimize}.}
\usage{
\S4method{summary}{StanOptimize}(object, ...)
}
\arguments{
\item{object}{A \code{StanOptimize} object.}

\item{...}{Additional arguments, currently unused.}
}
\description{
Summary method for objects of class \code{StanOptimize}.
}
