% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathfinder.R
\docType{methods}
\name{summary-StanPathfinder}
\alias{summary-StanPathfinder}
\alias{summary,StanPathfinder-method}
\title{Summary method for objects of class \code{StanPathfinder}.}
\usage{
\S4method{summary}{StanPathfinder}(object, ...)
}
\arguments{
\item{object}{A \code{StanPathfinder} object.}

\item{...}{Additional arguments, currently unused.}
}
\description{
Summary method for objects of class \code{StanPathfinder}.
}
