% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variational.R
\docType{methods}
\name{summary-StanVariational}
\alias{summary-StanVariational}
\alias{summary,StanVariational-method}
\title{Summary method for objects of class \code{StanVariational}.}
\usage{
\S4method{summary}{StanVariational}(object, ...)
}
\arguments{
\item{object}{A \code{StanVariational} object.}

\item{...}{Additional arguments, currently unused.}
}
\description{
Summary method for objects of class \code{StanVariational}.
}
