% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_methods.R
\docType{methods}
\name{unconstrain_draws}
\alias{unconstrain_draws}
\alias{unconstrain_draws,StanBase-method}
\alias{unconstrain_draws,StanBase,StanBase-method}
\alias{unconstrain_draws,StanOptimize-method}
\alias{unconstrain_draws,StanOptimize,StanOptimize-method}
\title{Unconstrain all parameter draws.}
\usage{
unconstrain_draws(stan_object, draws = NULL)

\S4method{unconstrain_draws}{StanBase}(stan_object, draws = NULL)

\S4method{unconstrain_draws}{StanOptimize}(stan_object, draws = NULL)
}
\arguments{
\item{stan_object}{A \code{StanBase} object.}

\item{draws}{(optional) A \verb{posterior::draws_*} object to be unconstrained
(instead of the draws in the \code{StanBase} object).}
}
\description{
Unconstrain all parameter draws.
}
