% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_methods.R
\docType{methods}
\name{unconstrain_variables}
\alias{unconstrain_variables}
\alias{unconstrain_variables,StanBase-method}
\alias{unconstrain_variables,StanBase,StanBase-method}
\title{Unconstrain a vector of variables.}
\usage{
unconstrain_variables(stan_object, variables)

\S4method{unconstrain_variables}{StanBase}(stan_object, variables)
}
\arguments{
\item{stan_object}{A \code{StanBase} object.}

\item{variables}{Vector of variables to be unconstrained.}
}
\description{
Unconstrain a vector of variables.
}
