% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{stan_diagnose}
\alias{stan_diagnose}
\title{stan_diagnose}
\usage{
stan_diagnose(
  fn,
  par_inits = NULL,
  n_pars = NULL,
  additional_args = list(),
  grad_fun = NULL,
  lower = -Inf,
  upper = Inf,
  eval_standalone = FALSE,
  globals = TRUE,
  packages = NULL,
  seed = NULL,
  refresh = NULL,
  quiet = FALSE,
  output_dir = NULL,
  output_basename = NULL,
  sig_figs = NULL
)
}
\arguments{
\item{fn}{Function to estimate parameters for}

\item{par_inits}{Initial values for parameters. This can either be a numeric vector
of initial values (which will be used for all chains), a list of numeric vectors (of length
equal to the number of chains), a function taking a single argument (the chain ID) and
returning a numeric vector of initial values, or NULL (in which case Stan will
generate initial values automatically).
(must be specified if \code{n_pars} is NULL)}

\item{n_pars}{Number of parameters to estimate
(must be specified if \code{par_inits} is NULL)}

\item{additional_args}{List of additional arguments to pass to the function}

\item{grad_fun}{Either:
\itemize{
\item \code{NULL} for finite-differences (default)
\item A function for calculating gradients w.r.t. each parameter
\item "RTMB" to use the RTMB package for automatic differentiation
}}

\item{lower}{Lower bound constraint(s) for parameters}

\item{upper}{Upper bound constraint(s) for parameters}

\item{eval_standalone}{(logical) Whether to evaluate the function in a
separate R session. Defaults to \code{FALSE}.}

\item{globals}{(optional) a logical, a character vector, or a named list
to control how globals are handled when evaluating functions in a
separate R session. Ignored if \code{eval_standalone} = \code{FALSE}.
For details, see section 'Globals used by future expressions'
in the help for \code{\link[future:future]{future::future()}}.}

\item{packages}{(optional) a character vector specifying packages
to be attached in the \R environment evaluating the function.
Ignored if \code{eval_standalone} = \code{FALSE}.}

\item{seed}{Random seed}

\item{refresh}{Number of iterations for printing}

\item{quiet}{(logical) Whether to suppress Stan's output}

\item{output_dir}{Directory to store outputs}

\item{output_basename}{Basename to use for output files}

\item{sig_figs}{Number of significant digits to use for printing}
}
\value{
\code{StanLaplace} object
}
\description{
Check gradient estimation using Stan's 'Diagnose' method
}
