% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{stan_diagnose}
\alias{stan_diagnose}
\title{stan_diagnose}
\usage{
stan_diagnose(
  fn,
  par_inits,
  additional_args = list(),
  grad_fun = NULL,
  lower = -Inf,
  upper = Inf,
  seed = NULL,
  refresh = NULL,
  quiet = FALSE,
  output_dir = NULL,
  output_basename = NULL,
  sig_figs = NULL
)
}
\arguments{
\item{fn}{Function to estimate parameters for}

\item{par_inits}{Initial values}

\item{additional_args}{List of additional arguments to pass to the function}

\item{grad_fun}{Function calculating gradients w.r.t. each parameter}

\item{lower}{Lower bound constraint(s) for parameters}

\item{upper}{Upper bound constraint(s) for parameters}

\item{seed}{Random seed}

\item{refresh}{Number of iterations for printing}

\item{quiet}{(logical) Whether to suppress Stan's output}

\item{output_dir}{Directory to store outputs}

\item{output_basename}{Basename to use for output files}

\item{sig_figs}{Number of significant digits to use for printing}
}
\value{
\code{StanLaplace} object
}
\description{
Check gradient estimation using Stan's 'Diagnose' method
}
