% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{stan_optimize}
\alias{stan_optimize}
\title{stan_optimize}
\usage{
stan_optimize(
  fn,
  par_inits,
  ...,
  algorithm = "lbfgs",
  grad_fun = NULL,
  lower = -Inf,
  upper = Inf,
  jacobian = FALSE,
  iter = 2000,
  save_iterations = FALSE,
  output_dir = tempdir(),
  control = list()
)
}
\description{
Estimate parameters using Stan's optimization algorithms
}
