% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{stan_optimize}
\alias{stan_optimize}
\title{stan_optimize}
\usage{
stan_optimize(
  fn,
  par_inits = NULL,
  n_pars = NULL,
  additional_args = list(),
  algorithm = "lbfgs",
  grad_fun = NULL,
  lower = -Inf,
  upper = Inf,
  eval_standalone = FALSE,
  globals = TRUE,
  packages = NULL,
  seed = NULL,
  refresh = NULL,
  quiet = FALSE,
  output_dir = NULL,
  output_basename = NULL,
  sig_figs = NULL,
  save_iterations = NULL,
  jacobian = NULL,
  init_alpha = NULL,
  iter = NULL,
  tol_obj = NULL,
  tol_rel_obj = NULL,
  tol_grad = NULL,
  tol_rel_grad = NULL,
  tol_param = NULL,
  history_size = NULL
)
}
\arguments{
\item{fn}{Function to estimate parameters for}

\item{par_inits}{Initial values for parameters
(must be specified if \code{n_pars} is NULL)}

\item{n_pars}{Number of parameters to estimate
(must be specified if \code{par_inits} is NULL)}

\item{additional_args}{List of additional arguments to pass to the function}

\item{algorithm}{(string) The optimization algorithm. One of \code{"lbfgs"},
\code{"bfgs"}, or \code{"newton"}.}

\item{grad_fun}{Function calculating gradients w.r.t. each parameter}

\item{lower}{Lower bound constraint(s) for parameters}

\item{upper}{Upper bound constraint(s) for parameters}

\item{eval_standalone}{(logical) Whether to evaluate the function in a
separate R session. Defaults to \code{FALSE}.}

\item{globals}{(optional) a logical, a character vector, or a named list
to control how globals are handled when evaluating functions in a
separate R session. Ignored if \code{eval_standalone} = \code{FALSE}.
For details, see section 'Globals used by future expressions'
in the help for \code{\link[future:future]{future::future()}}.}

\item{packages}{(optional) a character vector specifying packages
to be attached in the \R environment evaluating the function.
Ignored if \code{eval_standalone} = \code{FALSE}.}

\item{seed}{Random seed}

\item{refresh}{Number of iterations for printing}

\item{quiet}{(logical) Whether to suppress Stan's output}

\item{output_dir}{Directory to store outputs}

\item{output_basename}{Basename to use for output files}

\item{sig_figs}{Number of significant digits to use for printing}

\item{save_iterations}{Save optimization iterations to output file}

\item{jacobian}{(logical) Whether or not to use the Jacobian adjustment for
constrained variables. For historical reasons, the default is \code{FALSE}, meaning optimization
yields the (regularized) maximum likelihood estimate. Setting it to \code{TRUE}
yields the maximum a posteriori estimate.}

\item{init_alpha}{(positive real) The initial step size parameter.}

\item{iter}{(positive integer) The maximum number of iterations.}

\item{tol_obj}{(positive real) Convergence tolerance on changes in objective function value.}

\item{tol_rel_obj}{(positive real) Convergence tolerance on relative changes in objective function value.}

\item{tol_grad}{(positive real) Convergence tolerance on the norm of the gradient.}

\item{tol_rel_grad}{(positive real) Convergence tolerance on the relative norm of the gradient.}

\item{tol_param}{(positive real) Convergence tolerance on changes in parameter value.}

\item{history_size}{(positive integer) The size of the history used when
approximating the Hessian. Only available for L-BFGS.}
}
\value{
\code{StanOptimize} object
}
\description{
Estimate parameters using Stan's optimization algorithms
}
