% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathfinder.R
\name{stan_pathfinder}
\alias{stan_pathfinder}
\title{stan_pathfinder}
\usage{
stan_pathfinder(
  fn,
  par_inits,
  ...,
  num_psis_draws = 1000,
  num_paths = 4,
  save_single_paths = FALSE,
  max_lbfgs_iters = 1000,
  num_draws = 1000,
  num_elbo_draws = 25,
  output_dir = tempdir(),
  control = list()
)
}
\description{
Estimate parameters using Stan's pathfinder algorithm
}
